api ws

+++ Shared/Component/Health.h
@@ -19,7 +19,7 @@ struct rr_component_health
     RR_SERVER_ONLY(float secondary_damage;)
     RR_SERVER_ONLY(float damage_reduction;)
     RR_SERVER_ONLY(float poison;)
-    RR_SERVER_ONLY(uint32_t poison_ticks;)
+    RR_SERVER_ONLY(uint16_t poison_ticks;)
     EntityIdx parent_id;
     uint8_t flags;
     RR_SERVER_ONLY(uint8_t protocol_state;)
